support multiple data types relevant to discrete sequences
