+++ server/README.md
@@ -1,11 +1,21 @@
-# tkn-api-server
+# TKN Server
 
-A high-performance token processing server built with Bun, featuring real-time token sequence parsing, Neo4j database integration, and Prometheus metrics monitoring.
+A high-performance token processing server component of the TKN system. This server is responsible for receiving, processing, and storing token sequences in a Neo4j/Memgraph database.
+
+## Role in the TKN System
+
+The TKN Server serves as the backend for the TKN protocol implementation:
+
+- Receives and processes messages using the TKN binary protocol
+- Mines token sequences from data streams using a custom token mining algorithm
+- Stores processed tokens in Neo4j/Memgraph for analysis and retrieval
+- Provides real-time processing with WebSocket communication
+- Monitors performance metrics using Prometheus
 
 ## Features
 
 - Real-time token sequence processing and parsing
-- Neo4j database integration for token storage
+- Neo4j/Memgraph database integration for token storage
 - Prometheus metrics monitoring
 - WebSocket-based communication
 - Efficient token batch processing
@@ -68,10 +78,19 @@ The server exposes Prometheus metrics at `http://localhost:4000/metrics`. Key me
 
 The server consists of several key components:
 
-- **Observer**: Processes incoming data buffers and identifies token sequences
+- **TknServer**: Main server implementation that handles socket connections and message processing
+- **TknMiner**: Processes incoming data buffers and identifies token sequences
 - **SyncStream**: Handles token batch processing and database storage
+- **SymbolTable**: Manages token symbol resolution and mapping
 - **Metrics Server**: Provides Prometheus metrics endpoint
-- **WebSocket Server**: Manages real-time client connections
+
+## Core Files
+
+- `src/index.ts`: Entry point for the server
+- `src/lib/tkn-server.ts`: Main server implementation
+- `src/lib/tkn-miner.ts`: Token mining logic
+- `src/lib/sync-stream.ts`: Token batch processing
+- `src/lib/metrics-server.ts`: Prometheus metrics server
 
 ## License
