+++ server/src/lib/symbol-table/symbol-table.ts
@@ -5,14 +5,12 @@
 
 import type { CryptoHasher } from "bun";
 import { LRUCache } from "lru-cache";
-
-export type HashedValue = Uint8Array<ArrayBuffer>;
-
-export enum HashAlgorithm {
-  SHA256 = "sha256",
-  MURMUR3 = "murmur3",
-  CYRB53 = "cyrb53",
-}
+import {
+  HashAlgorithm,
+  type HashedValue,
+  murmurhash3_32,
+  cyrb53,
+} from "./hash-algorithms";
 
 export class SymbolTable {
   private table = new Map<string, any>();
@@ -85,84 +83,6 @@ export class SymbolTable {
   }
 
   /**
-   * Implementation of MurmurHash3 (32-bit variant)
-   * A fast non-cryptographic hash function
-   */
-  private murmurhash3_32(data: string, seed = 0): HashedValue {
-    let h = seed;
-    const k = 0x5bd1e995;
-    const r = 24;
-    const len = data.length;
-
-    // Process 4 bytes at a time
-    for (let i = 0; i < len; i += 4) {
-      let k1 =
-        data.charCodeAt(i & 0xff) |
-        (((i + 1 < len ? data.charCodeAt(i + 1) : 0) & 0xff) << 8) |
-        (((i + 2 < len ? data.charCodeAt(i + 2) : 0) & 0xff) << 16) |
-        (((i + 3 < len ? data.charCodeAt(i + 3) : 0) & 0xff) << 24);
-
-      k1 = Math.imul(k1, k);
-      k1 = (k1 << r) | (k1 >>> (32 - r));
-      k1 = Math.imul(k1, k);
-
-      h = Math.imul(h, k);
-      h ^= k1;
-    }
-
-    // Finalization
-    h ^= len;
-    h ^= h >>> 16;
-    h = Math.imul(h, 0x85ebca6b);
-    h ^= h >>> 13;
-    h = Math.imul(h, 0xc2b2ae35);
-    h ^= h >>> 16;
-
-    // Create HashedValue from the 32-bit integer
-    const result = new Uint8Array(this.hashSize);
-    for (let i = 0; i < Math.min(4, this.hashSize); i++) {
-      result[i] = (h >>> (i * 8)) & 0xff;
-    }
-
-    return result;
-  }
-
-  /**
-   * Implementation of cyrb53 hash
-   * A simple and effective 64-bit non-cryptographic hash function
-   */
-  private cyrb53(str: string, seed = 0): HashedValue {
-    let h1 = 0xdeadbeef ^ seed;
-    let h2 = 0x41c6ce57 ^ seed;
-
-    for (let i = 0; i < str.length; i++) {
-      const ch = str.charCodeAt(i);
-      h1 = Math.imul(h1 ^ ch, 2654435761);
-      h2 = Math.imul(h2 ^ ch, 1597334677);
-    }
-
-    h1 = Math.imul(h1 ^ (h1 >>> 16), 2246822507);
-    h1 ^= Math.imul(h2 ^ (h2 >>> 13), 3266489909);
-    h2 = Math.imul(h2 ^ (h2 >>> 16), 2246822507);
-    h2 ^= Math.imul(h1 ^ (h1 >>> 13), 3266489909);
-
-    // Create result buffer
-    const result = new Uint8Array(this.hashSize);
-
-    // Fill the first 8 bytes with our 64-bit result
-    const bytesToFill = Math.min(8, this.hashSize);
-    for (let i = 0; i < bytesToFill; i++) {
-      if (i < 4) {
-        result[i] = (h2 >>> (i * 8)) & 0xff;
-      } else {
-        result[i] = (h1 >>> ((i - 4) * 8)) & 0xff;
-      }
-    }
-
-    return result;
-  }
-
-  /**
    * Convert arbitrary data to a buffer hash using the selected algorithm
    * Uses object pooling and caching for efficiency
    */
@@ -181,10 +101,10 @@ export class SymbolTable {
     // Choose the hashing algorithm
     switch (this.algorithm) {
       case HashAlgorithm.MURMUR3:
-        hashArray = this.murmurhash3_32(inputStr);
+        hashArray = murmurhash3_32(inputStr, 0, this.hashSize);
         break;
       case HashAlgorithm.CYRB53:
-        hashArray = this.cyrb53(inputStr);
+        hashArray = cyrb53(inputStr, 0, this.hashSize);
         break;
       case HashAlgorithm.SHA256:
       default:
